--- src/app/dataview/energy/popover/energy-settings-popover.component.html
@@ -1,3 +1,0 @@
-<p>
-  energy-settings-popover works!
-</p>
