--- src/app/dataview/energy/popover/energy-settings-popover.component.spec.ts
@@ -1,25 +1,0 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { EnergySettingsPopoverComponent } from './energy-settings-popover.component';
-
-describe('EnergySettingsPopoverComponent', () => {
-  let component: EnergySettingsPopoverComponent;
-  let fixture: ComponentFixture<EnergySettingsPopoverComponent>;
-
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [ EnergySettingsPopoverComponent ]
-    })
-    .compileComponents();
-  }));
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(EnergySettingsPopoverComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
