--- src/app/dataview/energy/popover/energy-settings-popover.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-energy-settings-popover',
-  templateUrl: './energy-settings-popover.component.html',
-  styleUrls: ['./energy-settings-popover.component.scss']
-})
-export class EnergySettingsPopoverComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
